requirements.txt
discord.py==2.7.0
openpyxl==3.11.5
